refactor(itemstate): split store.go into project_apply.go and keys.go
Move applyProjectItem/applyShallowItem/applyProbeItem/stateFrom into
project_apply.go and the key/string utilities into keys.go, per ADR-010's
one-concern-per-file convention. Pure code motion, no behavior change.
