+++ internal/itemstate/store.go
@@ -3,8 +3,6 @@ package itemstate
 import (
 	"errors"
 	"reflect"
-	"strconv"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
@@ -821,128 +819,6 @@ func (s *Store) updateIndexes(before, after ItemState, key string) {
 
 // ---- helpers ----
 
-// applyProjectItem copies all fields from a gh.ProjectItem into item and returns ChangeFlags.
-func applyProjectItem(item *ItemState, pi gh.ProjectItem) ChangeFlags {
-	var flags ChangeFlags
-
-	if item.ID != pi.ID && pi.ID != "" {
-		item.ID = pi.ID
-	}
-	if item.ItemID != pi.ItemID {
-		item.ItemID = pi.ItemID
-	}
-
-	if item.Title != pi.Title || item.Body != pi.Body || item.URL != pi.URL || item.Author != pi.Author {
-		item.Title = pi.Title
-		item.Body = pi.Body
-		item.URL = pi.URL
-		item.Author = pi.Author
-		flags |= TitleBodyChanged
-	}
-
-	if !reflect.DeepEqual(item.Assignees, pi.Assignees) {
-		item.Assignees = copyStrings(pi.Assignees)
-		flags |= AssigneesChanged
-	}
-
-	if item.State != stateFrom(pi) || item.IsClosed != pi.IsClosed || item.IsPR != pi.IsPR {
-		item.State = stateFrom(pi)
-		item.IsClosed = pi.IsClosed
-		item.IsPR = pi.IsPR
-		flags |= StateChanged
-	}
-
-	if !reflect.DeepEqual(item.Labels, pi.Labels) {
-		item.Labels = copyStrings(pi.Labels)
-		flags |= LabelsChanged
-	}
-
-	if item.Status != pi.Status {
-		item.Status = pi.Status
-		flags |= StatusChanged
-	}
-
-	if !item.UpdatedAt.Equal(pi.UpdatedAt) {
-		item.UpdatedAt = pi.UpdatedAt
-	}
-
-	if !reflect.DeepEqual(item.BlockedBy, pi.BlockedBy) {
-		item.BlockedBy = copyDeps(pi.BlockedBy)
-		flags |= BlockedByChanged
-	}
-
-	if !reflect.DeepEqual(item.Comments, pi.Comments) {
-		item.Comments = copyComments(pi.Comments)
-		flags |= CommentsChanged
-	}
-
-	// Sync LinkedPR fields from ProjectItem.
-	if pi.LinkedPRNumber != 0 {
-		if item.LinkedPR == nil {
-			item.LinkedPR = &LinkedPRState{}
-			flags |= LinkedPRChanged
-		}
-		lpr := item.LinkedPR
-		if lpr.Number != pi.LinkedPRNumber ||
-			lpr.HeadSHA != "" || // already set by richer fetch — don't overwrite
-			!reflect.DeepEqual(lpr.Reviews, pi.LinkedPRReviews) ||
-			!reflect.DeepEqual(lpr.ReviewRequests, pi.LinkedPRReviewRequests) ||
-			!reflect.DeepEqual(lpr.ThreadComments, pi.LinkedPRReviewThreadComments) ||
-			lpr.ResolvedThreadCount != pi.LinkedPRResolvedThreadCount ||
-			lpr.IsMergeQueueEnabled != pi.LinkedPRIsMergeQueueEnabled ||
-			lpr.IsInMergeQueue != pi.LinkedPRIsInMergeQueue ||
-			!reflect.DeepEqual(lpr.MergeQueueEntry, pi.LinkedPRMergeQueueEntry) {
-			if lpr.Number != pi.LinkedPRNumber {
-				lpr.Number = pi.LinkedPRNumber
-				flags |= LinkedPRChanged
-			}
-			if !reflect.DeepEqual(lpr.Reviews, pi.LinkedPRReviews) {
-				lpr.Reviews = copyPRReviews(pi.LinkedPRReviews)
-				flags |= LinkedPRChanged
-			}
-			if !reflect.DeepEqual(lpr.ReviewRequests, pi.LinkedPRReviewRequests) {
-				lpr.ReviewRequests = copyReviewRequests(pi.LinkedPRReviewRequests)
-				flags |= LinkedPRChanged
-			}
-			if !reflect.DeepEqual(lpr.ThreadComments, pi.LinkedPRReviewThreadComments) {
-				lpr.ThreadComments = copyComments(pi.LinkedPRReviewThreadComments)
-				flags |= LinkedPRChanged | CommentsChanged
-			}
-			if lpr.ResolvedThreadCount != pi.LinkedPRResolvedThreadCount {
-				lpr.ResolvedThreadCount = pi.LinkedPRResolvedThreadCount
-				flags |= LinkedPRChanged
-			}
-			if lpr.IsMergeQueueEnabled != pi.LinkedPRIsMergeQueueEnabled {
-				lpr.IsMergeQueueEnabled = pi.LinkedPRIsMergeQueueEnabled
-				flags |= LinkedPRChanged
-			}
-			if lpr.IsInMergeQueue != pi.LinkedPRIsInMergeQueue {
-				lpr.IsInMergeQueue = pi.LinkedPRIsInMergeQueue
-				flags |= LinkedPRChanged
-			}
-			if !reflect.DeepEqual(lpr.MergeQueueEntry, pi.LinkedPRMergeQueueEntry) {
-				lpr.MergeQueueEntry = pi.LinkedPRMergeQueueEntry
-				flags |= LinkedPRChanged
-			}
-		}
-	} else if item.LinkedPR != nil && item.LinkedPR.Number != 0 {
-		// PR was delinked — only clear the number, preserve any richer state.
-		// (Full removal would need an explicit IssuePRDelinked mutation.)
-	}
-
-	if flags == 0 && item.Repo == "" {
-		// First-time population of identity fields.
-		item.Repo = pi.Repo
-		item.Number = pi.Number
-		return TitleBodyChanged | StatusChanged | LabelsChanged
-	}
-
-	item.Repo = pi.Repo
-	item.Number = pi.Number
-
-	return flags
-}
-
 // HasItems reports whether the Store contains at least one item. O(1) and
 // non-allocating. Safe to call concurrently with Apply.
 func (s *Store) HasItems() bool {
@@ -1125,103 +1001,6 @@ func (s *Store) Reset(items []gh.ProjectItem) {
 	}
 }
 
-// applyShallowItem updates only the shallow board fields of item from pi.
-// Deep fields (Body, Comments, Assignees, BlockedBy, LinkedPRReviews, etc.)
-// are left unchanged. Used by CacheImpl.Reconcile to apply shallow board
-// updates without wiping deep-fetched data.
-func applyShallowItem(item *ItemState, pi gh.ProjectItem) ChangeFlags {
-	var flags ChangeFlags
-
-	if item.ID != pi.ID && pi.ID != "" {
-		item.ID = pi.ID
-	}
-	if item.ItemID != pi.ItemID && pi.ItemID != "" {
-		item.ItemID = pi.ItemID
-	}
-
-	if item.Title != pi.Title {
-		item.Title = pi.Title
-		flags |= TitleBodyChanged
-	}
-
-	if item.URL != pi.URL && pi.URL != "" {
-		item.URL = pi.URL
-		flags |= TitleBodyChanged
-	}
-
-	if item.State != stateFrom(pi) || item.IsClosed != pi.IsClosed || item.IsPR != pi.IsPR {
-		item.State = stateFrom(pi)
-		item.IsClosed = pi.IsClosed
-		item.IsPR = pi.IsPR
-		flags |= StateChanged
-	}
-
-	if !reflect.DeepEqual(item.Labels, pi.Labels) {
-		item.Labels = copyStrings(pi.Labels)
-		flags |= LabelsChanged
-	}
-
-	if item.Status != pi.Status {
-		item.Status = pi.Status
-		item.Terminal = false
-		flags |= StatusChanged
-	}
-
-	if !item.UpdatedAt.Equal(pi.UpdatedAt) {
-		item.UpdatedAt = pi.UpdatedAt
-	}
-
-	return flags
-}
-
-// applyProbeItem updates only the probe-visible fields of item from pi.
-// Labels are explicitly NOT updated — the probe query fetches no label data.
-// Wiping Labels here would silently discard the cached label set.
-func applyProbeItem(item *ItemState, pi gh.BoardProbeItem) ChangeFlags {
-	var flags ChangeFlags
-
-	if item.ID != pi.ContentID && pi.ContentID != "" {
-		item.ID = pi.ContentID
-	}
-	if item.ItemID != pi.ItemID && pi.ItemID != "" {
-		item.ItemID = pi.ItemID
-	}
-
-	probeState := "open"
-	if pi.IsClosed {
-		probeState = "closed"
-	}
-	if item.State != probeState || item.IsClosed != pi.IsClosed || item.IsPR != pi.IsPR {
-		item.State = probeState
-		item.IsClosed = pi.IsClosed
-		item.IsPR = pi.IsPR
-		flags |= StateChanged
-	}
-
-	if item.Status != pi.Status {
-		item.Status = pi.Status
-		item.Terminal = false
-		flags |= StatusChanged
-	}
-
-	if !item.UpdatedAt.Equal(pi.EffectiveUpdatedAt) {
-		item.UpdatedAt = pi.EffectiveUpdatedAt
-	}
-
-	// Labels are intentionally not touched here. The probe query contains no
-	// label data; setting item.Labels from an empty probe would wipe the
-	// cached label set populated by FetchItemDetails or the bootstrap query.
-
-	return flags
-}
-
-func stateFrom(pi gh.ProjectItem) string {
-	if pi.IsClosed {
-		return "closed"
-	}
-	return "open"
-}
-
 func ensureLinkedPR(item *ItemState, prNumber int) {
 	if item.LinkedPR == nil {
 		item.LinkedPR = &LinkedPRState{Number: prNumber}
@@ -1262,74 +1041,3 @@ func ensureStageStateMaps(item *ItemState) {
 	}
 }
 
-func containsString(ss []string, s string) bool {
-	for _, v := range ss {
-		if v == s {
-			return true
-		}
-	}
-	return false
-}
-
-func removeReviewRequest(rrs []gh.ReviewRequest, login string) []gh.ReviewRequest {
-	out := rrs[:0:0]
-	for _, rr := range rrs {
-		if rr.Login != login {
-			out = append(out, rr)
-		}
-	}
-	if len(out) == 0 {
-		return nil
-	}
-	return out
-}
-
-// unionStrings returns a slice containing all elements of existing plus any
-// elements from incoming that are not already present. Order is preserved.
-func unionStrings(existing, incoming []string) []string {
-	result := copyStrings(existing)
-	for _, s := range incoming {
-		if !containsString(result, s) {
-			result = append(result, s)
-		}
-	}
-	return result
-}
-
-func removeString(ss []string, s string) []string {
-	out := ss[:0:0]
-	for _, v := range ss {
-		if v != s {
-			out = append(out, v)
-		}
-	}
-	if len(out) == 0 {
-		return nil
-	}
-	return out
-}
-
-// prKeyFor constructs the PR reverse-index key "owner/repo#pr<N>".
-// Same semantics as boardcache.prKey; duplicated here to keep the Store
-// package self-contained (boardcache is a separate package).
-func prKeyFor(repo string, prNum int) string {
-	return repo + "#pr" + strconv.Itoa(prNum)
-}
-
-// parseKey splits "owner/repo#N" into repo and number.
-func parseKey(key string) (string, int) {
-	idx := strings.LastIndex(key, "#")
-	if idx < 0 {
-		return key, 0
-	}
-	repo := key[:idx]
-	numStr := key[idx+1:]
-	n := 0
-	for _, c := range numStr {
-		if c < '0' || c > '9' {
-			return repo, 0
-		}
-		n = n*10 + int(c-'0')
-	}
-	return repo, n
-}
